+++ src/Controller/TestController.php
@@ -3,9 +3,13 @@
 
 namespace App\Controller;
 
+use App\Entity\Educator;
 use App\Entity\Test;
+use App\Enums\Image;
 use App\Form\CreateType\TestCreateType;
 use App\Form\UpdateType\TestUpdateType;
+use App\Util\Helper;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -29,20 +33,35 @@ class TestController extends AbstractController
      */
     public function create(Request $request)
     {
+        $user = $this->getUser();
+        if (!$user instanceof Educator) {
+            $this->addFlash('success', 'У вас нет доступа к этой странице.');
+            return new RedirectResponse($this->generateUrl('page.home'));
+        }
         $test = new Test();
+        $test->setCreator($this->getUser());
 
         $form = $this->createForm(TestCreateType::class, $test);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($test);
-            $em->flush();
+            $files = $request->files->get('test_create');
+            $file = Helper::uploadFile(array_shift($files), $this->getParameter('upload_dir'));
+            if ($file)
+            {
+                $test->setImage($file);
 
-            $this->addFlash('success', "Тест успешно создан.");
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($test);
+                $em->flush();
 
-            return new RedirectResponse($this->generateUrl('page.home'));
+                $this->addFlash('success', 'Тест успешно создан.');
+
+                return new RedirectResponse($this->generateUrl('page.home'));
+            } else {
+                $this->addFlash('fail', 'Прикрепление изображения обязательно.');
+            }
         }
 
         return $this->render('forms/test/test.html.twig', [
@@ -60,6 +79,11 @@ class TestController extends AbstractController
      */
     public function update(Test $test, Request $request)
     {
+        $user = $this->getUser();
+        if (!$user instanceof Educator) {
+            $this->addFlash('success', 'У вас нет доступа к этой странице.');
+            return new RedirectResponse($this->generateUrl('page.home'));
+        }
         $form = $this->createForm(TestUpdateType::class, $test);
 
         $form->handleRequest($request);
@@ -69,6 +93,11 @@ class TestController extends AbstractController
         }
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $files = $request->files->get('test_update');
+            $file = Helper::uploadFile(array_shift($files), $this->getParameter('upload_dir'));
+            if ($file) {
+                $test->setImage($file);
+            }
             $em = $this->getDoctrine()->getManager();
             $em->persist($test);
             $em->flush();
@@ -93,6 +122,10 @@ class TestController extends AbstractController
      */
     public function delete(Test $test, Request $request)
     {
+        $user = $this->getUser();
+        if (!$user instanceof Educator) {
+            return new RedirectResponse($this->generateUrl('page.home'));
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($test);
         $em->flush();
